change lua initialization parameters and program names
